% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreenaProgramme.R
\name{AgreenaProgramme}
\alias{AgreenaProgramme}
\title{Agreena RothC}
\usage{
AgreenaProgramme(
  lonlat,
  sim_period,
  inp_s = "same_as_base",
  cp_b = c("spring", "winter", "none", "catch"),
  cp_s = c("spring", "winter", "none", "catch"),
  till_b = c("Reduced tillage", "Conventional tillage", "No tillage", "Not available"),
  till_s = c("Reduced tillage", "Conventional tillage", "No tillage", "Not available"),
  soiltype = "Clay",
  soil_data = "lucas"
)
}
\arguments{
\item{lonlat}{Longitude and latitude vector}

\item{sim_period}{Simulating period}

\item{inp_s}{Biomass inputs in the test scenario}

\item{cp_b}{Logical vector with months where the soil is covered. If only a
single value is passed the whole period will be treated homogeneously.}

\item{cp_s}{Logical vector with months where the soil is covered. If only a
single value is passed the whole period will be treated homogeneously.}

\item{till_b}{date ranges}

\item{till_s}{date ranges}

\item{soiltype}{argument necessary if \code{soil_data} is numerical}

\item{soil_data}{Either "lucas", "isric" or a number describing SOM in \%}

\item{inp_b}{Biomass inputs in the baseline scenario}
}
\value{
returns a data frame with average (AV) and standard deviation (SD)
of surface temperature (TS), bias corrected precipitation (PRECTOTCORR) and
Evapotranspiration Energy Flux (EVPTRNS).
}
\description{
Calculates soil carbon stocks for the RothC spin up period,
baseline and scenario runs.
}
\details{
Agreena RothC initializes RothC biomass inputs by inverse modeling
to calibrate total soil carbon stocks with estimates generated by ISRIC
soilgrids. The estimates biomass inputs are used to spin up the model until
soil carbon stocks equilibrium is reached for each soil carbon compartment.
Those values are then used to initialize Baseline and scenario runs.
}
\author{
Marcos Alves
}
