% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_moist_rothc.R
\name{f_moist_rothc}
\alias{f_moist_rothc}
\title{Effect of moisture on C flows}
\usage{
f_moist_rothc(pp, et, s_thick, pclay, pE = 1, soil_cover = TRUE)
}
\arguments{
\item{pp}{A vector with monthly precipitation (mm).}

\item{et}{A vector with same length with open pan evaporation or
evapotranspiration (mm).}

\item{s_thick}{Soil thickness in cm. Default for Rothamsted is 23 cm.}

\item{pE}{Evaporation coefficient. If open pan evaporation is used pE=0.75.
If Potential evaporation is used, pE=1.0.}

\item{soil_cover}{Logical vector with months where the soil is covered. If only a single value is passed the whole period will be treated homogeneously.}

\item{pClay}{Percent clay.}
}
\value{
A vector with the rate modifying factor.
}
\description{
Extended version of \code{\link[SoilR]{fW.RothC}} coping with different soil cover regimes.
}
\details{
Effects of moisture on decomposition rates according to the RothC model

Calculates the effects of moisture (precipitation and pan evaporation) on
decomposition rates according to the RothC model.
}
\author{
Marcos Alves
}
