% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_isric_soil_profile_rothc.R
\name{get_isric_soil_profile_rothc}
\alias{get_isric_soil_profile_rothc}
\title{Download a soil profile from the ISRIC soil database}
\usage{
get_isric_soil_profile_rothc(
  lonlat,
  statistic = c("mean", "Q0.5"),
  find.location.name = TRUE
)
}
\arguments{
\item{lonlat}{Longitude and latitude vector (e.g. c(-93, 42)).}

\item{statistic}{default is the mean}

\item{find.location.name}{default is TRUE. Use either maps package or photon API to find Country/State.
If you are running this function many times it might be better to set this to FALSE.}
}
\value{
data frame with soil characteristics and \code{meta} Attributes.
}
\description{
Retrieves soil data from the ISRIC global database and converts it to a data frame that can be used to run RothC
}
\details{
Source: https://www.isric.org/ \cr
Details: https://www.isric.org/explore/soilgrids/faq-soilgrids \cr

Variable which are directly retrieved and a simple unit conversion is performed: \cr
\itemize{
\item Bulk density - bdod \cr
\item Carbon - ocs \cr
\item Clay - clay \cr
\item Sand - sand \cr
\item Silt - silt \cr
\code{attr(x, "meta")} loads meta information about the location from where the soil profile was created.
}
}
\examples{
\dontrun{
## Get soil profile properties for a single point
sp1 <- get_isric_soil_profile_rothc(lonlat = c(-93, 42))
sp1
attr(sp1, "meta")
}

}
\seealso{
\code{\link{apsimx_soil_profile}}, \code{\link{edit_apsim_replace_soil_profile}}, \code{\link{edit_apsimx_replace_soil_profile}}.
}
\author{
Marcos Alves (Modified from: Fernando E. Miguez)
}
