% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lucas_soil_profile_rothc.R
\name{get_lucas_soil_profile_rothc}
\alias{get_lucas_soil_profile_rothc}
\title{Download a soil profile from the Lucas soil database}
\usage{
get_lucas_soil_profile_rothc(lonlat)
}
\arguments{
\item{lonlat}{Longitude and latitude vector (e.g. c(-93, 42)).}
}
\value{
data frame with soil characteristics and \code{meta} Attributes.
}
\description{
Retrieves soil data from the lucas database and converts it to a data frame that can be used to run RothC
}
\details{
Source:https://esdac.jrc.ec.europa.eu/projects/lucas/ \cr

Variable which are directly retrieved: \cr
\itemize{
\item Bulk density - BD (g/cm3)\cr
\item Carbon - SOC (tC/ha) \cr
\item Clay - ParticleSizeClay (\%)\cr
}
}
\examples{
\dontrun{
## Get soil profile properties for a single point
sp1 <- get_isric_soil_profile_rothc(lonlat = c(-93, 42))
sp1
attr(sp1, "meta")
}
}
\author{
Marcos Alves
}
