% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgreenaRothC.R, R/get_wth_power_nasa.R
\name{get_wth_power_nasa}
\alias{get_wth_power_nasa}
\alias{AgreenaRothC}
\title{Agreena RothC}
\usage{
AgreenaRothC(
  lonlat,
  sim_period,
  inp_b = 1.7,
  inp_s = 1.7,
  cp_b = FALSE,
  cp_s = TRUE,
  till_b = c("Till", "noTill", "redTill"),
  till_s = c("Till", "noTill", "redTill")
)

get_wth_power_nasa(lonlat, dates)
}
\arguments{
\item{lonlat}{Longitude and latitude vector}

\item{sim_period}{Simulating period}

\item{inp_b}{Biomass inputs in the baseline scenario}

\item{inp_s}{Biomass inputs in the test scenario}

\item{cp_b}{Logical vector with months where the soil is covered. If only a single value is passed the whole period will be treated homogeneously.}

\item{cp_s}{date ranges}

\item{till_b}{date ranges}

\item{till_s}{date ranges}

\item{dates}{Data range defining the period that should be used to create the \href{https://en.wikipedia.org/wiki/Climatological_normal}{"Climatolocal Normals"} for the specified locations.}
}
\value{
returns a data frame with average (AV) and standard deviation (SD) of surface temperature (TS), bias corrected precipitation (PRECTOTCORR) and Evapotranspiration Energy Flux (EVPTRNS).

returns a data frame with average (AV) and standard deviation (SD) of surface temperature (TS), bias corrected precipitation (PRECTOTCORR) and Evapotranspiration Energy Flux (EVPTRNS).
}
\description{
Calculates soil carbon stocks for the RothC spin up period, baseline and scenario runs.

Uses \code{\link[nasapower]{get_power}} from the \CRANpkg{nasapower} package to download data and calculate yearly means and standard deviation of temperature, precipitation and evapotranspiration.
}
\details{
Agreena RothC initializes RothC biomass inputs by inverse modeling to calibrate total soil carbon stocks with estimates generated by ISRIC soilgrids.
The estimates biomass inputs are used to spin up the model until soil carbon stocks equilibrium is reached for each soil carbon compartment. Those values are then used to initialize Baseline and scenario runs.

This function requires the \CRANpkg{nasapower} package version 4.0.0.

A summary description of nasa power variables can be found here: https://gist.github.com/abelcallejo/d68e70f43ffa1c8c9f6b5e93010704b8
TODO: Check the unit measures of the different outputs
}
\examples{
\dontrun{
require(nasapower)
pwr <- get_wth_power_nasa(lonlat = c(-93,42), dates = c("2012-01-01","2012-12-31"))
}

}
\author{
Marcos Alves
}
