% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rothc.R
\name{plot_rothc}
\alias{plot_rothc}
\title{Plot AgreenaRothC diagnostics plots}
\usage{
plot_rothc(...)
}
\arguments{
\item{...}{list of objects output from \link{AgreenaRothC}}

\item{x}{Results to be plotted (either "scenario" or "baseline")}
}
\value{
Long formatted soil carbon values that can be ploted by \link{plot_rothc}
}
\description{
Plot AgreenaRothC diagnostics plots
}
\details{
Receives as inputs a list of data frames containing the output of AgreenaRothC and plots diagnostics graphs comparing different runs
}
\seealso{
\link{AgreenaRothC}
}
\author{
Marcos Alves
}
