% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRothC.R
\name{plotRothC}
\alias{plotRothC}
\title{Plot AgreenaRothC diagnostics plots}
\usage{
plotRothC(x = "scenario", ...)
}
\arguments{
\item{x}{Results to be plotted (either "scenario" or "baseline")}

\item{...}{list of objects output from \link{AgreenaRothC}}
}
\value{
Long formatted soil carbon values that can be ploted by \link{plotRothC}
}
\description{
Plot AgreenaRothC diagnostics plots
}
\details{
Receives as inputs a list of data frames containing the output of AgreenaRothC and plots diagnostics graphs comparing different runs
}
\seealso{
\link{AgreenaRothC}
}
\author{
Marcos Alves
}
