% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWth.R
\name{plotWth}
\alias{plotWth}
\title{Plot weather climate Normals}
\usage{
plotWth(wth)
}
\arguments{
\item{wth}{data output from \link{get_wth_power_nasa}}
}
\value{
Long formatted soil carbon values that can be ploted by \link{plotRothC}
}
\description{
Plot weather climate Normals
}
\details{
Plots weather data Climate normals download from nasa power
}
\seealso{
\link{get_wth_power_nasa}
}
\author{
Marcos Alves
}
