% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyRothC.R
\name{tidyRothC}
\alias{tidyRothC}
\title{Tidy AgreenaRothC outputs}
\usage{
tidyRothC(...)
}
\arguments{
\item{...}{list of objects output from \link{AgreenaRothC}}
}
\value{
Long formatted soil carbon values that can be ploted by \link{plotRothC}
}
\description{
Auxiliary function to transform RothC individual results in tidy long data frames.
}
\details{
Receives as inputs a list of data frames containing the output of AgreenaRothC and clean the data by keeping only the soil carbon stocks values and tranforming it from wide to long with a names for each datafram inputed.
}
\seealso{
\link{AgreenaRothC}, \link{plotRothC}
}
\author{
Marcos Alves
}
