% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_to_resid.R
\name{yield_to_resid}
\alias{yield_to_resid}
\title{Yield to residues}
\usage{
yield_to_resid(yield, cropname)
}
\arguments{
\item{cropname}{vector with cropnames}

\item{yiled}{vector with crop yields}
}
\value{
vector with crop residues in t C/ha
}
\description{
Calculate crop specific biomass residues
baseline and scenario runs.
}
\author{
Marcos Alves
}
