% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wth_power_nasa.R
\name{get_wth_power_nasa}
\alias{get_wth_power_nasa}
\title{Download NASA-POWER weather data}
\usage{
get_wth_power_nasa(lonlat, dates)
}
\arguments{
\item{lonlat}{Longitude and latitude vector}

\item{dates}{Data range defining the period that should be used to create the \href{https://en.wikipedia.org/wiki/Climatological_normal}{"Climatolocal Normals"} for the specified locations.}
}
\value{
returns a data frame with average (AV) and standard deviation (SD) of surface temperature (TS), bias corrected precipitation (PRECTOTCORR) and Evapotranspiration Energy Flux (EVPTRNS).
}
\description{
Uses \code{\link[nasapower]{get_power}} from the \CRANpkg{nasapower} package to download data and calculate yearly means and standard deviation of temperature, precipitation and evapotranspiration.
}
\details{
This function requires the \CRANpkg{nasapower} package version 4.0.0.

A summary description of nasa power variables can be found here: https://gist.github.com/abelcallejo/d68e70f43ffa1c8c9f6b5e93010704b8
TODO: Check the unit measures of the different outputs
}
\examples{
\dontrun{
require(nasapower)
pwr <- get_wth_power_nasa(lonlat = c(-93, 42), dates = c("2012-01-01", "2012-12-31"))
}

}
\author{
Marcos Alves
}
