% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wth.R
\name{plot_wth}
\alias{plot_wth}
\title{Plot weather climate Normals}
\usage{
plot_wth(wth)
}
\arguments{
\item{wth}{data output from \link{get_wth_power_nasa}}
}
\value{
Long formatted soil carbon values that can be ploted by \link{plot_rothc}
}
\description{
Plot weather climate Normals
}
\details{
Plots weather data Climate normals download from nasa power
}
\seealso{
\link{get_wth_power_nasa}
}
\author{
Marcos Alves
}
